create delete modal
show a popup when delete button clicked

+++ hewan/index.php
@@ -61,7 +61,7 @@ require '../connect.php'; // Ensure this file exists and the path is correct
 
                             echo '<td>
                             <button><a href="update.php?hewanId=' . $id . '">Edit</a></button>
-                            <button><a href="delete.php?hewanId=' . $id . '">Delete</a></button>
+                            <button><a href="delete.php?hewanId=' . $id . '" onclick="return confirm(\'Apakah kamu yakin ingin menghapus item ini?\');">Delete</a></button>
                             </td>' .
                                 '</tr>';
                         }
