+++ orang/index.php
@@ -65,7 +65,7 @@ require '../connect.php'; // Ensure this file exists and the path is correct
 
                             echo '<td>
                             <button><a href="update.php?orangId=' . $id . '">Edit</a></button>
-                            <button><a href="delete.php?orangId=' . $id . '">Delete</a></button>
+                            <button><a href="delete.php?orangId=' . $id . '" onclick="return confirm(\'Apakah kamu yakin ingin menghapus item ini?\');">Delete</a></button>
                             </td>' .
                                 '</tr>';
                         }
